refactor(UserName): 사용자 이름 원시값 포장
